\name{loss}
\alias{gaussloss}
\alias{gaussngra}
\alias{hingeloss}
\alias{hingengra}
\alias{loss}
\alias{ngradient}
\alias{plotCVbst}
\alias{cvfolds}
\alias{error.bars}
\alias{loss.mhingebst}
\alias{mhingebst_fit}
\alias{balanced.folds}
\alias{permute.rows}
\title{Internal Function}
\description{Internal Function}
