\name{bst-package}
\alias{bst-package}
\docType{package}
\title{
Boosting for Classification and Regression
}
\description{
Gradient descent boosting for hinge loss and square error loss.
}
\details{
\tabular{ll}{
Package: \tab bst\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2010-04-15\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Zhu Wang
}
